Extract podcast conversion to the model itself
This is not exposed publicly.
